lint: added comments for new methods
